+++ src/candlewick/multibody/Visualizer.h
@@ -56,7 +56,7 @@ class Visualizer final : public BaseVisualizer {
   bool m_showGui = true;
   bool m_shouldExit = false;
   entt::entity m_grid, m_triad;
-  RobotDebugSystem *robotDebug = nullptr;
+  RobotDebugSystem *m_robotDebug = nullptr;
   std::vector<entt::entity> m_debug_frame_pos;
   std::vector<entt::entity> m_debug_frame_vel;