+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- bindings/python : add expose-debug.cpp, expose DebugMeshComponent
+
 ### Changed
 
 - slightly rework file dialog GUI
+- core/DebugScene : make setupPipelines() private, do not use optional
+- multibody/Visualizer/GUI : display info for robot debug frames
 
 ### Fixed